Implement Modern Search Widget with Enhanced Features
- Added SearchInputWidget for modern search input with animated placeholder and clear button.
- Integrated FilterPanelWidget for advanced filtering options including categories, tags, and properties.
- Implemented real-time search suggestions and search history management.
- Enhanced UI with theme-aware styling and smooth animations.
- Added keyboard shortcuts and accessibility support.
- Established signal-slot connections for search and filter interactions.
- Included necessary header files and forward declarations for widget components.

+++ examples/gallery/src/ui/widgets/viewers/ImageViewerWidget.h
@@ -166,7 +166,7 @@ public:
     // Getters
     QString getCurrentImage() const { return m_currentImage; }
     int getCurrentIndex() const { return m_currentIndex; }
-    int getImageCount() const { return m_imageList.size(); }
+    int getImageCount() const { return static_cast<int>(m_imageList.size()); }
     ViewMode getViewMode() const { return m_viewMode; }
     double getZoomFactor() const;
