+++ tests/test_performance_regression.h
@@ -1,6 +1,6 @@
 /**
  * QtLucide Performance Regression Tests - Header
- * 
+ *
  * Comprehensive benchmarking tests to prevent performance degradation
  * across icon rendering, memory usage, search operations, and application lifecycle.
  */
@@ -40,7 +40,7 @@ struct PerformanceMetrics {
     qint64 totalTime = 0;
     int iterations = 0;
     double standardDeviation = 0.0;
-    
+
     void addMeasurement(qint64 time) {
         totalTime += time;
         iterations++;
@@ -48,16 +48,16 @@ struct PerformanceMetrics {
         if (time > maxTime) maxTime = time;
         averageTime = totalTime / iterations;
     }
-    
+
     void calculateStandardDeviation(const QList<qint64>& measurements) {
         if (measurements.size() < 2) return;
-        
+
         double sum = 0.0;
         for (qint64 time : measurements) {
-            double diff = time - averageTime;
+            double diff = static_cast<double>(time) - static_cast<double>(averageTime);
             sum += diff * diff;
         }
-        standardDeviation = sqrt(sum / (measurements.size() - 1));
+        standardDeviation = sqrt(sum / (static_cast<double>(measurements.size()) - 1));
     }
 };
 
@@ -128,36 +128,36 @@ private:
     qint64 measureSearchTime(const QString& searchTerm, int iterations);
     qint64 measureApplicationStartup();
     qint64 measureApplicationShutdown();
-    
+
     // Benchmark helpers
     void warmupSystem();
     void clearSystemCaches();
     QList<qint64> runBenchmark(std::function<qint64()> benchmark, int iterations);
     bool isPerformanceAcceptable(const PerformanceMetrics& metrics, qint64 threshold);
-    
+
     // Memory tracking helpers
     qint64 getCurrentMemoryUsage();
     void trackMemoryAllocations(bool enable);
-    
+
     // Test data and objects
     lucide::QtLucide* m_lucide;
     IconMetadataManager* m_metadataManager;
     SearchWidget* m_searchWidget;
     IconGridWidget* m_iconGrid;
     QWidget* m_testWidget;
-    
+
     // Performance tracking
     QElapsedTimer m_timer;
     QList<PerformanceMetrics> m_performanceHistory;
     QList<MemoryMetrics> m_memoryHistory;
-    
+
     // Baseline performance thresholds (in milliseconds)
     static constexpr qint64 ICON_RENDER_THRESHOLD_MS = 5;
     static constexpr qint64 SEARCH_RESPONSE_THRESHOLD_MS = 100;
     static constexpr qint64 STARTUP_THRESHOLD_MS = 2000;
     static constexpr qint64 SHUTDOWN_THRESHOLD_MS = 1000;
     static constexpr qint64 MEMORY_LEAK_THRESHOLD_KB = 1024;
-    
+
     // Test configuration
     static constexpr int PERFORMANCE_ITERATIONS = 100;
     static constexpr int STRESS_TEST_ITERATIONS = 1000;
